Move search endpoint functionality to be an option in the opportunity endpoint
