+++ server/api/opportunity/__tests__/opportunity.fixture.js
@@ -13,7 +13,7 @@ const opList = [
   },
   {
     title: '2 Self driving model cars',
-    subtitle: 'using algorithmns to follow lines and avoid obstacles',
+    subtitle: 'using algorithms to follow lines and avoid obstacles',
     imgUrl: 'http://www.plaz-tech.com/wp-content/plugins/wp-easycart-data/products/pics1/Arduino%20Car%202_8ab5dd38f1e3f6f05ad244f1e5e74529.jpg',
     description: '# NZTA Innovation Centre\n \n We have 6 model cars with sensors for vision, proximity etc, \n controlled by Arduinos teach them to solve \n 4 challenges - move, follow a line, avoid obstacles, \n get to a destination etc. \n \n ## We need:\n * Open space with room for the test tracks - e.g a school hall\n * teams of 5 students\n * on adult helper per team, should be able to follow instructions and understand a little C++\n \n ## Learning outcomes:\n * programming a remote device\n * simple coding\n * algorithmic thinking\n * problem solving.\n \n',
     duration: '4 hours',
