--- server/api/search/__tests__/search.fixture.js
@@ -1,39 +1,0 @@
-const opList = [
-  {
-    title: '1 Mentor a year 12 business Impact Project',
-    subtitle: 'Help us create a programme connecting business with senior students',
-    description: 'We want to set up our Connect Ed programme, help us understand how to communication with businesses, build a website and contact group etc.',
-    offerOrg: 'Albany High School',
-    reqestor: 'Arthur Henry Smith',
-    location: 'Albany High School, Auckland',
-    status: 'active'
-    // tags: ['C++','Year 12','Business']
-  },
-  {
-    title: '2 Self driving model cars',
-    subtitle: 'using algorithms to follow lines and avoid obstacles',
-    description: '# NZTA Innovation Centre\n \n We have 6 model cars with sensors for vision, proximity etc, \n controlled by Arduinos teach them to solve \n 4 challenges - move, follow a line, avoid obstacles, \n get to a destination etc. \n \n ## We need:\n * Open space with room for the test tracks - e.g a school hall\n * teams of 5 students\n * on adult helper per team, should be able to follow instructions and understand a little C++\n \n ## Learning outcomes:\n * programming a remote device\n * simple coding\n * algorithmic thinking\n * problem solving.\n \n',
-    location: 'NZTA Innovation Centre, 5 Cook St Auckland',
-    status: 'active'
-    // tags: ['Robots','Algorithms']
-  },
-  {
-    title: '3 Growing in the garden',
-    subtitle: 'Growing digitally in the garden',
-    description: 'Project to grow something in the garden',
-    location: 'Newmarket, Auckland',
-    status: 'draft'
-    // tags: ['Garden']
-  },
-  {
-    title: '4 The first 100 metres',
-    subtitle: 'Launching into space',
-    description: 'Project to build a simple rocket that will reach 100m',
-    location: 'Albany, Auckland',
-    status: 'draft'
-    // tags: ['Rockets']
-  }
-
-]
-
-module.exports = opList
