--- server/api/search/search.routes.js
@@ -1,5 +1,0 @@
-const { searchOpportunities } = require('./search.controller')
-
-module.exports = function (server) {
-  server.get('/api/search', searchOpportunities)
-}
